backend/requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
cfgv==3.3.1 ; python_full_version >= '3.6.1'
distlib==0.3.6
exceptiongroup==1.1.0 ; python_version < '3.11'
faker==16.8.1
filelock==3.9.0 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
identify==2.5.17 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
nodeenv==1.7.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.0 ; python_version >= '3.7'
platformdirs==3.0.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pre-commit==3.0.4
pytest==7.2.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0 ; python_version >= '3.6'
rfc3986[idna2008]==1.5.0
setuptools==67.2.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
tomli==2.0.1 ; python_version < '3.11'
virtualenv==20.19.0 ; python_version >= '3.7'
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
bert-extractive-summarizer==0.10.1
blis==0.7.9
cachetools==5.3.0 ; python_version ~= '3.7'
catalogue==2.0.8 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
cloud-sql-python-connector[pymysql]==1.1.0
confection==0.0.4 ; python_version >= '3.6'
cryptography==39.0.1 ; python_version >= '3.6'
cymem==2.0.7
fastapi==0.91.0
frozenlist==1.3.3 ; python_version >= '3.7'
google-auth==2.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
huggingface-hub==0.12.0 ; python_full_version >= '3.7.0'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
keybert==0.7.0
langcodes==3.3.0 ; python_version >= '3.6'
markdown-it-py==2.1.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
murmurhash==1.0.9 ; python_version >= '3.6'
nltk==3.8.1 ; python_version >= '3.7'
numpy==1.24.2 ; python_version >= '3.8'
pathy==0.10.1 ; python_version >= '3.6'
pillow==9.4.0 ; python_version >= '3.7'
preshed==3.0.8 ; python_version >= '3.6'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4 ; python_version >= '3.7'
pygments==2.14.0 ; python_version >= '3.6'
pymysql==1.0.2
python-dotenv==0.21.1
regex==2022.10.31 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rich==13.3.1 ; python_full_version >= '3.7.0'
rsa==4.9 ; python_version >= '3.6'
scikit-learn==1.2.1 ; python_version >= '3.8'
scipy==1.10.0 ; python_version < '3.12' and python_version >= '3.8'
sentence-transformers==2.2.2
sentencepiece==0.1.97
smart-open==6.3.0 ; python_version >= '3.6' and python_version < '4.0'
spacy==3.5.0
spacy-legacy==3.0.12 ; python_version >= '3.6'
spacy-loggers==1.0.4 ; python_version >= '3.6'
sqlalchemy==2.0.3
srsly==2.4.5 ; python_version >= '3.6'
starlette==0.24.0 ; python_version >= '3.7'
thinc==8.1.7 ; python_version >= '3.6'
threadpoolctl==3.1.0 ; python_version >= '3.6'
tokenizers==0.13.2
torch==1.13.1 ; python_full_version >= '3.7.0'
torchvision==0.14.1 ; python_version >= '3.7'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==4.26.1 ; python_full_version >= '3.7.0'
typer==0.7.0 ; python_version >= '3.6'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.20.0
wasabi==1.1.1 ; python_version >= '3.6'
yarl==1.8.2 ; python_version >= '3.7'
